Fix Last-Modified header
The PUT header was incorrectly set and the convertion to RFC822
date format failed.
Instead of using time.time(), if metadata were available, we now
use mtime properties.

The GET header was using ctime instead of mtime : in case of
overwritten object, it was the timestamp of first upload
that was send.

The use of Timestamp class was also removed since none of its features
was used to `Last Modified` header.

Fix OS-351