Optimize bucket listing
When Container Hierarchy is used, a lot of containers
may be created. This patch will use first character of
encoded delimiter of Container Hierarchy as delimiter.

Since S3 doesn't support delimiter in bucket listing (and
results are not compliant with bucket name for old SDS),
we also skip `subdir` entries when generating results.